+++ src/dflow/io.py
@@ -61,6 +61,14 @@ class InputArtifacts(AutonamedDict):
             else:
                 self.template.inputs.parameters["dflow_%s_path_list" % key] = \
                     InputParameter(value=[])
+        if config["lineage"] and self.template is not None and \
+                key[:6] != "dflow_":
+            if isinstance(value.source, S3Artifact):
+                self.template.inputs.parameters["dflow_%s_urn" % key] = \
+                    InputParameter(value=value.source.urn)
+            else:
+                self.template.inputs.parameters["dflow_%s_urn" % key] = \
+                    InputParameter(value="")
 
     def set_template(self, template):
         super().set_template(template)
@@ -74,6 +82,16 @@ class InputArtifacts(AutonamedDict):
                     self.template.inputs.parameters["dflow_%s_path_list"
                                                     % name] = InputParameter(
                         value=[])
+        if config["lineage"]:
+            for name, art in self.items():
+                if name[:6] == "dflow_":
+                    continue
+                if isinstance(art.source, S3Artifact):
+                    self.template.inputs.parameters["dflow_%s_urn" % name] = \
+                        InputParameter(value=art.source.urn)
+                else:
+                    self.template.inputs.parameters["dflow_%s_urn" % name] = \
+                        InputParameter(value="")
 
 
 class OutputParameters(AutonamedDict):
@@ -90,6 +108,10 @@ class OutputArtifacts(AutonamedDict):
             self.template.outputs.parameters["dflow_%s_path_list" % key] = \
                 OutputParameter(value=[])
             value.handle_path_list()
+        if config["lineage"] and self.template is not None:
+            self.template.outputs.parameters["dflow_%s_urn" % key] = \
+                OutputParameter(value="")
+            value.handle_urn()
 
     def set_template(self, template):
         super().set_template(template)
@@ -98,6 +120,11 @@ class OutputArtifacts(AutonamedDict):
                 self.template.outputs.parameters["dflow_%s_path_list" % name]\
                     = OutputParameter(value=[])
                 art.handle_path_list()
+        if config["lineage"]:
+            for name, art in self.items():
+                self.template.outputs.parameters["dflow_%s_urn" % name]\
+                    = OutputParameter(value="")
+                art.handle_urn()
 
 
 class ArgoVar:
@@ -387,6 +414,7 @@ class InputArtifact(ArgoVar):
         optional: optional artifact or not
         type: artifact type
         source: default source
+        archive: regarded as archived file or not
     """
 
     def __init__(
@@ -401,7 +429,7 @@ class InputArtifact(ArgoVar):
                           "OutputArtifact", S3Artifact] = None,
             mode: Optional[int] = None,
             sub_path: Optional[str] = None,
-            **kwargs,
+            archive: str = "default",
     ) -> None:
         self.path = path
         self.name = name
@@ -413,6 +441,7 @@ class InputArtifact(ArgoVar):
         self._sub_path = None
         self.mode = mode
         self.sub_path = sub_path
+        self.archive = archive
 
     def __getattr__(self, key):
         from .task import Task
@@ -451,35 +480,44 @@ class InputArtifact(ArgoVar):
         return self.template.inputs.parameters["dflow_%s_path_list" %
                                                self.name]
 
+    def get_urn_parameter(self):
+        return self.template.inputs.parameters["dflow_%s_urn" % self.name]
+
     def convert_to_argo(self):
+        archive = None
+        if self.archive is None:
+            archive = V1alpha1ArchiveStrategy(_none={})
         if self.path in NotAllowedInputArtifactPath:
             raise RuntimeError(
                 "Path [%s] is not allowed for input artifact" % self.path)
         if self.source is None:
             return V1alpha1Artifact(name=self.name, path=self.path,
-                                    optional=self.optional, mode=self.mode)
+                                    optional=self.optional, mode=self.mode,
+                                    archive=archive)
         if isinstance(self.source, (InputArtifact, OutputArtifact)):
             sub_path = self.sub_path if self.sub_path is not None else \
                 self.source._sub_path
             return V1alpha1Artifact(name=self.name, path=self.path,
                                     optional=self.optional,
                                     _from=str(self.source), sub_path=sub_path,
-                                    mode=self.mode)
+                                    mode=self.mode, archive=archive)
         elif isinstance(self.source, S3Artifact):
             if s3_config["repo_type"] == "s3":
                 return V1alpha1Artifact(name=self.name, path=self.path,
                                         optional=self.optional, s3=self.source,
-                                        sub_path=self.sub_path, mode=self.mode)
+                                        sub_path=self.sub_path, mode=self.mode,
+                                        archive=archive)
             else:
                 return V1alpha1Artifact(name=self.name, path=self.path,
                                         optional=self.optional,
                                         oss=self.source.oss(),
-                                        sub_path=self.sub_path, mode=self.mode)
+                                        sub_path=self.sub_path, mode=self.mode,
+                                        archive=archive)
         elif isinstance(self.source, str):
             return V1alpha1Artifact(name=self.name, path=self.path,
                                     optional=self.optional,
                                     raw=V1alpha1RawArtifact(data=self.source),
-                                    mode=self.mode)
+                                    mode=self.mode, archive=archive)
         else:
             raise RuntimeError(
                 "Cannot pass an object of type %s to artifact %s" %
@@ -707,7 +745,6 @@ class OutputArtifact(ArgoVar):
             archive: str = "default",
             global_name: Optional[str] = None,
             from_expression: Union[IfExpression, str] = None,
-            **kwargs,
     ) -> None:
         self.path = path
         self.name = name
@@ -757,6 +794,8 @@ class OutputArtifact(ArgoVar):
         if config["save_path_as_parameter"] and key in ["_from",
                                                         "from_expression"]:
             self.handle_path_list()
+        if config["lineage"] and key in ["_from", "from_expression"]:
+            self.handle_urn()
 
     def handle_path_list(self):
         if self.template is not None:
@@ -778,8 +817,31 @@ class OutputArtifact(ArgoVar):
                     _else=self.from_expression._else.get_path_list_parameter(),
                 )
 
+    def handle_urn(self):
+        if self.template is not None:
+            if hasattr(self, "_from") and self._from is not None:
+                self.template.outputs.parameters["dflow_%s_urn" %
+                                                 self.name].default = ""
+                self.template.outputs.parameters[
+                    "dflow_%s_urn" % self.name].value_from_parameter = \
+                    self._from.get_urn_parameter()
+            elif hasattr(self, "from_expression") and self.from_expression is \
+                    not None:
+                self.template.outputs.parameters["dflow_%s_urn" %
+                                                 self.name].default = ""
+                self.template.outputs.parameters[
+                    "dflow_%s_urn" % self.name].value_from_expression = \
+                    if_expression(
+                    _if=self.from_expression._if,
+                    _then=self.from_expression._then.get_urn_parameter(),
+                    _else=self.from_expression._else.get_urn_parameter(),
+                )
+
     def get_path_list_parameter(self):
         return self.step.outputs.parameters["dflow_%s_path_list" % self.name]
+
+    def get_urn_parameter(self):
+        return self.step.outputs.parameters["dflow_%s_urn" % self.name]
 
     def __repr__(self):
         from .task import Task