+++ src/dflow/python/python_op_template.py
@@ -1,5 +1,4 @@
 import inspect
-import json
 import os
 import random
 import string
@@ -44,6 +43,7 @@ class Slices:
         pool_size: for multi slices per step, use a multiprocessing pool to
             handle each slice, 1 for serial, -1 for infinity (i.e. equals to
             the number of slices)
+        register_first_only: only register first slice when lineage used
     """
 
     def __init__(
@@ -55,6 +55,7 @@ class Slices:
             output_artifact: Optional[List[str]] = None,
             sub_path: bool = False,
             pool_size: Optional[int] = None,
+            register_first_only: bool = False,
     ) -> None:
         self.input_parameter = input_parameter if input_parameter is not \
             None else []
@@ -72,6 +73,7 @@ class Slices:
         else:
             self.slices = "{{item}}"
         self.pool_size = pool_size
+        self.register_first_only = register_first_only
 
 
 def get_source_code(o):
@@ -153,7 +155,6 @@ class PythonOPTemplate(PythonScriptOPTemplate):
                  envs: Dict[str, str] = None,
                  init_containers: Optional[List[V1alpha1UserContainer]] = None,
                  tmp_root: str = "/tmp",
-                 **kwargs,
                  ) -> None:
         self.n_parts = {}
         self.op_class = op_class
@@ -193,7 +194,8 @@ class PythonOPTemplate(PythonScriptOPTemplate):
             if isinstance(sign, Artifact):
                 self.inputs.artifacts[name] = InputArtifact(
                     path=r"%s/inputs/artifacts/" % self.tmp_root + name,
-                    optional=sign.optional, type=sign.type)
+                    optional=sign.optional, type=sign.type,
+                    archive=sign.archive)
             elif isinstance(sign, BigParameter) and config["mode"] != "debug":
                 self.inputs.parameters[name] = InputParameter(
                     save_as_artifact=True, path=r"%s/inputs/parameters/"
@@ -219,6 +221,11 @@ class PythonOPTemplate(PythonScriptOPTemplate):
                         value_from_path=r"%s/outputs/parameters/"
                         "dflow_%s_path_list" % (self.tmp_root, name),
                         default=[])
+                if config["lineage"]:
+                    self.outputs.parameters["dflow_%s_urn" % name] = \
+                        OutputParameter(
+                        value_from_path="%s/outputs/parameters/dflow_%s_urn"
+                        % (self.tmp_root, name), default="")
             elif isinstance(sign, BigParameter) and config["mode"] != "debug":
                 self.outputs.parameters[name] = OutputParameter(
                     save_as_artifact=True,
@@ -374,7 +381,8 @@ class PythonOPTemplate(PythonScriptOPTemplate):
 
         script += "import json, jsonpickle\n"
         script += "from dflow import config, s3_config\n"
-        script += "config.update(json.loads('%s'))\n" % json.dumps(config)
+        script += "config.update(jsonpickle.loads('%s'))\n" % \
+            jsonpickle.dumps(config)
         script += "s3_config.update(jsonpickle.loads('%s'))\n" % \
             jsonpickle.dumps(s3_config)
         if op_class.__module__ in ["__main__", "__mp_main__"]:
@@ -404,7 +412,7 @@ class PythonOPTemplate(PythonScriptOPTemplate):
         script += "from dflow.python.utils import handle_input_artifact," \
                   " handle_input_parameter\n"
         script += "from dflow.python.utils import handle_output_artifact," \
-                  " handle_output_parameter\n"
+                  " handle_output_parameter, handle_lineage\n"
         script += f"from {op_class.__module__} import {class_name}\n\n"
         script += "if __name__ == '__main__':\n"
         if hasattr(op_class, "func"):
@@ -523,6 +531,30 @@ class PythonOPTemplate(PythonScriptOPTemplate):
                 script += "    handle_output_parameter('%s', output['%s'], "\
                     "output_sign['%s'], %s, r'%s')\n" % (name, name, name,
                                                          slices, self.tmp_root)
+        if config["lineage"]:
+            if self.slices is not None and self.slices.register_first_only:
+                if "{{item}}" in self.dflow_vars:
+                    var_name = self.dflow_vars["{{item}}"]
+                else:
+                    var_name = "dflow_var_%s" % len(self.dflow_vars)
+                    self.inputs.parameters[var_name] = InputParameter(
+                        value="{{item}}")
+                    self.dflow_vars["{{item}}"] = var_name
+                self.inputs.parameters["dflow_first"] = InputParameter()
+                if not hasattr(self, "first_var"):
+                    self.first_var = "r'''{{inputs.parameters.dflow_first}}'''"
+                script += "    if r'''{{inputs.parameters.%s}}''' == "\
+                    "%s:\n" % (var_name, self.first_var)
+            else:
+                script += "    if True:\n"
+            script += "        input_urns = {}\n"
+            for name, sign in input_sign.items():
+                if isinstance(sign, Artifact):
+                    script += "        input_urns['%s'] = '{{inputs."\
+                        "parameters.dflow_%s_urn}}'\n" % (name, name)
+            script += "        handle_lineage('{{workflow.name}}', "\
+                "'{{pod.name}}', op_obj, input_urns, '{{workflow.parameters."\
+                "dflow_workflow_urn}}', '%s')\n" % self.tmp_root
 
         self.script = script